backend/requirements.txt
fastapi>=0.125.0
uvicorn[standard]>=0.38.0
sqlmodel>=0.0.27
asyncpg>=0.31.0
psycopg2-binary>=2.9.11
python-dotenv>=1.0.0
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.21
passlib==1.7.4
bcrypt==3.2.0
Pillow>=10.0.0
transformers>=4.30.0
torch>=2.0.0
